GBE-110: Moved a few models out of remains. Wasn't able to get Act, unfortunately
