+++ expo/gbe/models/__init__.py
@@ -1,7 +1,7 @@
-from user_message import UserMessage
-from available_interest import AvailableInterest
 from remains import (
     Conference,
+    ConferenceDay,
+    VolunteerWindow,
     Biddable,
     Profile,
     Performer,
@@ -14,22 +14,20 @@ from remains import (
     TechInfo,
     CueInfo,
     Act,
-    Room,
-    ConferenceDay,
-    VolunteerWindow,
     Event,
     Show,
-    GenericEvent,
     Class,
-    BidEvaluation,
-    PerformerFestivals,
-    Volunteer,
-    Vendor,
-    AdBid,
-    ArtBid,
     Costume,
-    ClassProposal,
-    ConferenceVolunteer,
-    ProfilePreferences
+    BidEvaluation,
+    Vendor,
+    Volunteer,
 )
+from available_interest import AvailableInterest
+from class_proposal import ClassProposal
+from conference_volunteer import ConferenceVolunteer
+from generic_event import GenericEvent
+from performer_festivals import PerformerFestivals
+from profile_preferences import ProfilePreferences
+from room import Room
+from user_message import UserMessage
 from volunteer_interest import VolunteerInterest
